feat(keymap): change layer holds from momentary to toggle
This is to prevent having to hold the thumb modifier key down to use
a layer.

During often repeated keys such as arrow movement, the thumb would hold
down the modifier key for a while and seemed to lead to strain for me.

Instead, the thumb modifier key needs to be pressed to toggle the layer,
and then manually toggled off.

It's an extra keypress, FWIW.

See: https://getreuer.info/posts/keyboards/thumb-ergo/index.html